refactor: move find chat platform to utils
