+++ libs/journeys/ui/src/components/Button/utils/findChatPlatform/findChatPlatform.ts
@@ -1,5 +1,5 @@
-import { MessagePlatform } from '../../../../__generated__/globalTypes'
-import { ButtonFields_action } from '../__generated__/ButtonFields'
+import { MessagePlatform } from '../../../../../__generated__/globalTypes'
+import { ButtonFields_action } from '../../__generated__/ButtonFields'
 
 export function findChatPlatform(
   buttonAction: ButtonFields_action | null
